【修改描述】过载控制优化，cleancode 【DTS/CR/AR单号】 AR20240127801698 【合入版本】BeiMing 24.2.T1.B001,
Match-id-a3775f08ed2deee73ceebc8ee65b89ffe5da8509

+++ src/cache/read/rcache.cpp
@@ -324,13 +324,13 @@ BResult RCache::GetSliceFromChunk(RCacheTierType tier, const RCacheChunkPtr &chu
 void RCache::GetCacheResource(uint64_t &memCap, uint64_t &memUsed, uint64_t &diskCap, uint64_t &diskUsed)
 {
     auto config = BioConfig::Instance()->GetDaemonConfig();
-    memCap = (config.memReadRatio * config.memCap) / NO_10;
+    memCap = (static_cast<uint64_t>(config.memReadRatio) * config.memCap) / NO_10;
     memUsed = FlowManager::GetCacheUsedSize(FLOW_RCACHE, FLOW_MEMORY);
     diskCap = 0;
     for (auto &item : config.diskCaps) {
         diskCap += static_cast<uint64_t>(item);
     }
-    diskCap = diskCap * config.diskReadRatio / NO_10;
+    diskCap = diskCap * static_cast<uint64_t>(config.diskReadRatio) / NO_10;
     diskUsed = FlowManager::GetCacheUsedSize(FLOW_RCACHE, FLOW_DISK);
 }
